# Optimize String.codepoints